+++ packages/core/src/context/provider.ts
@@ -1,4 +1,4 @@
-import type { ProgramOptionsKey } from '../tasks/constants';
+import type { ProgramOptionsKey } from '../tasks/taskOptions';
 import type { RnvContext } from './types';
 
 // We separate this context access from the rest to avoid circular reference issues
